[FE] feat: #57 payment backend <-> frontend API 연동 코드 구현 및 payment 로직 분리
payment 관련 API가 늘어나서 따로 파일로 분리했습니다.
